feat: NumPy 2.4.2 alignment investigation with bug fixes
This commit addresses findings from a comprehensive NumPy alignment
investigation, achieving 92% compatibility (84/91 tests pass).

## Bug Fixes

### Task #84: Empty array max/min throws ArgumentException
- `np.max([])` and `np.min([])` now throw ArgumentException
- Matches NumPy's "zero-size array to reduction operation" ValueError
- Files: Default.Reduction.AMax.cs, Default.Reduction.AMin.cs

### Task #85: argmax/argmin return Int64 (not Int32)
- Changed return type from int to long to match NumPy
- Updated IL kernel wrapper to use Convert.ToInt64() for boxed results
- Fixed test files expecting int to use long
- Files: Default.Reduction.ArgMax/Min.cs, NDArray.argmax/argmin.cs,
         DefaultEngine.ReductionOp.cs, np.argmax.cs

## New Features

### Task #86: np.nanmean, np.nanstd, np.nanvar
- Implements NaN-ignoring statistical functions
- Full axis support with keepdims parameter
- ddof parameter for std/var (delta degrees of freedom)
- Files: np.nanmean.cs, np.nanstd.cs, np.nanvar.cs

## Test Coverage

- TypePromotionTests.cs: 34 tests for NEP50 type promotion rules
- EmptyArrayReductionTests.cs: Empty array edge case validation
- Updated ArgMaxNaNTests.cs, NDArray.AMax.Test.cs for long return type

## Documentation

- docs/NUMPY_ALIGNMENT_INVESTIGATION.md: Comprehensive NumPy 2.4.2
  behavior reference and test results by category

Investigation tasks completed: #80-#86 (7 tasks)
Overall alignment: 92% (84/91 tests pass)

+++ src/NumSharp.Core/Backends/Default/Math/DefaultEngine.ReductionOp.cs
@@ -221,19 +221,20 @@ namespace NumSharp.Backends
         /// Returns the index of the maximum value.
         /// </summary>
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        protected int argmax_elementwise_il(NDArray arr)
+        protected long argmax_elementwise_il(NDArray arr)
         {
             if (arr.Shape.IsScalar || (arr.Shape.NDim == 1 && arr.Shape.size == 1))
                 return 0;
 
             var inputType = arr.GetTypeCode;
 
-            // ArgMax always returns int, but needs accumulator type for comparison
+            // ArgMax returns long (int64) to match NumPy 2.x behavior
+            // Internally uses int kernels (arrays rarely exceed 2^31 elements), widens to long for API
             // For floating point types, use scalar implementation which handles NaN correctly (NumPy: first NaN wins)
             // For Boolean, use scalar implementation (IL doesn't support bool comparison directly)
             return inputType switch
             {
-                NPTypeCode.Boolean => (int)argmax_elementwise(arr),  // Boolean scalar path
+                NPTypeCode.Boolean => (long)(int)argmax_elementwise(arr),  // Boolean scalar path
                 NPTypeCode.Byte => ExecuteElementReduction<int>(arr, ReductionOp.ArgMax, NPTypeCode.Byte),
                 NPTypeCode.Int16 => ExecuteElementReduction<int>(arr, ReductionOp.ArgMax, NPTypeCode.Int16),
                 NPTypeCode.UInt16 => ExecuteElementReduction<int>(arr, ReductionOp.ArgMax, NPTypeCode.UInt16),
@@ -241,8 +242,8 @@ namespace NumSharp.Backends
                 NPTypeCode.UInt32 => ExecuteElementReduction<int>(arr, ReductionOp.ArgMax, NPTypeCode.UInt32),
                 NPTypeCode.Int64 => ExecuteElementReduction<int>(arr, ReductionOp.ArgMax, NPTypeCode.Int64),
                 NPTypeCode.UInt64 => ExecuteElementReduction<int>(arr, ReductionOp.ArgMax, NPTypeCode.UInt64),
-                NPTypeCode.Single => (int)argmax_elementwise(arr),  // NaN-aware scalar path
-                NPTypeCode.Double => (int)argmax_elementwise(arr),  // NaN-aware scalar path
+                NPTypeCode.Single => (long)(int)argmax_elementwise(arr),  // NaN-aware scalar path
+                NPTypeCode.Double => (long)(int)argmax_elementwise(arr),  // NaN-aware scalar path
                 NPTypeCode.Decimal => ExecuteElementReduction<int>(arr, ReductionOp.ArgMax, NPTypeCode.Decimal),
                 _ => throw new NotSupportedException($"ArgMax not supported for type {inputType}")
             };
@@ -253,19 +254,20 @@ namespace NumSharp.Backends
         /// Returns the index of the minimum value.
         /// </summary>
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        protected int argmin_elementwise_il(NDArray arr)
+        protected long argmin_elementwise_il(NDArray arr)
         {
             if (arr.Shape.IsScalar || (arr.Shape.NDim == 1 && arr.Shape.size == 1))
                 return 0;
 
             var inputType = arr.GetTypeCode;
 
-            // ArgMin always returns int, but needs accumulator type for comparison
+            // ArgMin returns long (int64) to match NumPy 2.x behavior
+            // Internally uses int kernels (arrays rarely exceed 2^31 elements), widens to long for API
             // For floating point types, use scalar implementation which handles NaN correctly (NumPy: first NaN wins)
             // For Boolean, use scalar implementation (IL doesn't support bool comparison directly)
             return inputType switch
             {
-                NPTypeCode.Boolean => (int)argmin_elementwise(arr),  // Boolean scalar path
+                NPTypeCode.Boolean => (long)(int)argmin_elementwise(arr),  // Boolean scalar path
                 NPTypeCode.Byte => ExecuteElementReduction<int>(arr, ReductionOp.ArgMin, NPTypeCode.Byte),
                 NPTypeCode.Int16 => ExecuteElementReduction<int>(arr, ReductionOp.ArgMin, NPTypeCode.Int16),
                 NPTypeCode.UInt16 => ExecuteElementReduction<int>(arr, ReductionOp.ArgMin, NPTypeCode.UInt16),
@@ -273,8 +275,8 @@ namespace NumSharp.Backends
                 NPTypeCode.UInt32 => ExecuteElementReduction<int>(arr, ReductionOp.ArgMin, NPTypeCode.UInt32),
                 NPTypeCode.Int64 => ExecuteElementReduction<int>(arr, ReductionOp.ArgMin, NPTypeCode.Int64),
                 NPTypeCode.UInt64 => ExecuteElementReduction<int>(arr, ReductionOp.ArgMin, NPTypeCode.UInt64),
-                NPTypeCode.Single => (int)argmin_elementwise(arr),  // NaN-aware scalar path
-                NPTypeCode.Double => (int)argmin_elementwise(arr),  // NaN-aware scalar path
+                NPTypeCode.Single => (long)(int)argmin_elementwise(arr),  // NaN-aware scalar path
+                NPTypeCode.Double => (long)(int)argmin_elementwise(arr),  // NaN-aware scalar path
                 NPTypeCode.Decimal => ExecuteElementReduction<int>(arr, ReductionOp.ArgMin, NPTypeCode.Decimal),
                 _ => throw new NotSupportedException($"ArgMin not supported for type {inputType}")
             };
